# specs/001-i-want-the/contracts/system-prompts.openapi.json
{
  "openapi": "3.0.1",
  "info": { "title": "System Prompts API", "version": "0.1.0" },
  "paths": {
    "/v1/system-prompts": {
      "get": {
        "summary": "List built-in and custom prompts",
        "responses": {
          "200": {"description": "OK", "content": {"application/json": {"schema": {"type": "object", "properties": {"built_ins": {"type": "array", "items": {"$ref": "#/components/schemas/BuiltInPrompt"}}, "custom": {"type": "array", "items": {"$ref": "#/components/schemas/CustomPrompt"}}}, "required": ["built_ins", "custom"] }}}}
        }
      },
      "post": {
        "summary": "Create custom prompt",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreatePromptInput"}}}},
        "responses": {"201": {"description": "Created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CustomPrompt"}}}}, "400": {"description": "Validation error"}}
      }
    },
    "/v1/system-prompts/{id}": {
      "patch": {"summary": "Update custom prompt","parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdatePromptInput"}}}}, "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CustomPrompt"}}}}, "400": {"description": "Bad request"}, "404": {"description": "Not found"}}},
      "delete": {"summary": "Delete custom prompt","parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}], "responses": {"204": {"description": "Deleted"}, "404": {"description": "Not found"}}}
    },
    "/v1/system-prompts/{id}/duplicate": {
      "post": {"summary": "Duplicate built-in or custom prompt","parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}], "responses": {"201": {"description": "Created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CustomPrompt"}}}}, "404": {"description": "Not found"}}}
    },
    "/v1/system-prompts/{id}/select": {
      "post": {"summary": "Select prompt for a conversation","parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SelectPromptInput"}}}}, "responses": {"200": {"description": "OK", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConversationSelectionResult"}}}}, "404": {"description": "Not found"}}}
    },
    "/v1/system-prompts/none/select": {
      "post": {"summary": "Clear active prompt for conversation","requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ClearSelectionInput"}}}}, "responses": {"200": {"description": "OK"}}}
    }
  },
  "components": {
    "schemas": {
      "BuiltInPrompt": {"type": "object", "properties": {"id": {"type": "string"}, "slug": {"type": "string"}, "name": {"type": "string"}, "description": {"type": "string"}, "order": {"type": "integer"}, "body": {"type": "string"}, "read_only": {"type": "boolean"}}, "required": ["id", "slug", "name", "order", "body", "read_only"]},
      "CustomPrompt": {"type": "object", "properties": {"id": {"type": "string"}, "name": {"type": "string"}, "body": {"type": "string"}, "usage_count": {"type": "integer"}, "last_used_at": {"type": ["string", "null"], "format": "date-time"}, "created_at": {"type": "string", "format": "date-time"}, "updated_at": {"type": "string", "format": "date-time"}}, "required": ["id", "name", "body", "usage_count", "created_at", "updated_at"]},
      "CreatePromptInput": {"type": "object", "properties": {"name": {"type": "string", "maxLength": 255}, "body": {"type": "string"}}, "required": ["name", "body"]},
      "UpdatePromptInput": {"type": "object", "properties": {"name": {"type": "string", "maxLength": 255}, "body": {"type": "string"}}, "minProperties": 1},
      "SelectPromptInput": {"type": "object", "properties": {"conversation_id": {"type": "string"}, "inline_override": {"type": ["string", "null"]}}, "required": ["conversation_id"]},
      "ClearSelectionInput": {"type": "object", "properties": {"conversation_id": {"type": "string"}}, "required": ["conversation_id"]},
      "ConversationSelectionResult": {"type": "object", "properties": {"conversation_id": {"type": "string"}, "active_system_prompt_id": {"type": ["string", "null"]}}, "required": ["conversation_id", "active_system_prompt_id"]}
    }
  }
}